Dev watch (#40)
* Put a glucose chart on the watch.

This incorporates a ton of fine work by Eric Jensen, but I've squashed
it down into a single commit for ease of management.

The watch now has two discrete pages, one which is the standard actions
based interface, the second is an information interface which shows the
IOB, COB, basal rate and a glucose chart.

Glucose data is read from HealthKit on the watch. Sometimes this lags
behind so we also make requests to backfill glucose data as
necessary. We're also minimizing the amount of data sent over BLE as
much as possible since BLE is slow and less reliable.

* Fix filename mismatches in comments